# Search v1: Remove unnecessary subqueries