--- a/pkg/services/sqlstore/permissions/dashboard_filter_no_subquery.go
+++ b/pkg/services/sqlstore/permissions/dashboard_filter_no_subquery.go
@@ -1,0 +1,5 @@
+package permissions
+
+type accessControlDashboardPermissionFilterNoFolderSubquery struct {
+	accessControlDashboardPermissionFilter
+}