<!--

Thank you for sending a pull request! Here are some tips:

1. If this is your first time, please read our contribution guide at https://github.com/grafana/grafana/blob/main/CONTRIBUTING.md

2. Ensure you include and run the appropriate tests as part of your Pull Request.

3. In a new feature or configuration option, an update to the documentation is necessary. Everything related to the documentation is under the docs folder in the root of the repository.

4. If the Pull Request is a work in progress, make use of GitHub's "Draft PR" feature and mark it as such.

5. If you can not merge your Pull Request due to a merge conflict, Rebase it. This gets it in sync with the main branch.

6. Name your PR as "<FeatureArea>: Describe your change", e.g. Alerting: Prevent race condition. If it's a fix or feature relevant for the changelog describe the user impact in the title. The PR title is used to auto-generate the changelog for issues marked with the "add to changelog" label.

7. If your PR content should be added to the What's New document for the next major or minor release, add the **add to what's new** label to your PR. Note that you should add this label to the main PR that introduces the feature; do not add this label to smaller PRs for the feature.

-->

**What is this feature?**

launch changes in #71253 behind a feature flag

- adds feature toggle `permissionsFilterRemoveSubquery`
- introduces a `PermissionsFilter` interface that `accessControlDashboardPermissionFilter` satisfies

- modifies `NewAccessControlDashboardPermissionFilter` constructor to call the right implementation depending on the feature flag
- modify tests for triggering both `accessControlDashboardPermissionFilter` and `accessControlDashboardPermissionFilterNoFolderSubquery` implementations
- adds an alternative `accessControlDashboardPermissionFilterNoFolderSubquery` implementation that replaces the subquery for fetching the dashboard folder with a join

Considering the query for a request like `GET /api/search?type=dash-db&query=dashboard_0_0`:

<details><summary>Existing query example</summary>
<p>

```
SELECT
                        dashboard.id,
                        dashboard.uid,
                        dashboard.title,
                        dashboard.slug,
                        dashboard_tag.term,
                        dashboard.is_folder,
                        dashboard.folder_id,
                        folder.uid AS folder_uid,
                        folder.slug AS folder_slug,
                        folder.title AS folder_title  FROM ( SELECT dashboard.id FROM dashboard WHERE ((dashboard.uid IN (SELECT substr(scope, 16) FROM permission WHERE scope LIKE 'dashboards:uid:%' AND role_id IN(SELECT id FROM role INNER JOIN (
                        SELECT ur.role_id
                        FROM user_role AS ur
                        WHERE ur.user_id = 1
                        AND (ur.org_id = 1 OR ur.org_id = 0)
                UNION
                        SELECT tr.role_id FROM team_role as tr
                        WHERE tr.team_id IN(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50)
                        AND tr.org_id = 1
                UNION
                        SELECT br.role_id FROM builtin_role AS br
                        WHERE br.role IN ("")
                        AND (br.org_id = 1 OR br.org_id = 0)
                ) as all_role ON role.id = all_role.role_id)  AND action = "dashboards:read") AND NOT dashboard.is_folder) OR (dashboard.folder_id IN (SELECT d.id FROM dashboard as d WHERE d.uid IN (SELECT substr(scope, 13) FROM permission WHERE scope LIKE 'folders:uid:%'  AND role_id IN(SELECT id FROM role INNER JOIN (
                        SELECT ur.role_id
                        FROM user_role AS ur
                        WHERE ur.user_id = 1
                        AND (ur.org_id = 1 OR ur.org_id = 0)
                UNION
                        SELECT tr.role_id FROM team_role as tr
                        WHERE tr.team_id IN(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50)
                        AND tr.org_id = 1
                UNION
                        SELECT br.role_id FROM builtin_role AS br
                        WHERE br.role IN ("")
                        AND (br.org_id = 1 OR br.org_id = 0)
                ) as all_role ON role.id = all_role.role_id)  AND action = "dashboards:read")) AND NOT dashboard.is_folder)) AND dashboard.org_id=1 AND dashboard.title LIKE "%dashboard_0_0%" AND dashboard.is_folder = 0 ORDER BY dashboard.title ASC LIMIT 1000 OFFSET 0) AS ids
                INNER JOIN dashboard ON ids.id = dashboard.id
LEFT OUTER JOIN dashboard AS folder ON folder.id = dashboard.folder_id
                LEFT OUTER JOIN dashboard_tag ON dashboard.id = dashboard_tag.dashboard_id
 ORDER BY dashboard.title ASC
```

</p>
</details> 

The modified query looks like:
<details><summary>New query example</summary>
<p>

```
SELECT
                        dashboard.id,
                        dashboard.uid,
                        dashboard.title,
                        dashboard.slug,
                        dashboard_tag.term,
                        dashboard.is_folder,
                        dashboard.folder_id,
                        folder.uid AS folder_uid,
                        folder.slug AS folder_slug,
                        folder.title AS folder_title  FROM ( SELECT dashboard.id FROM dashboard LEFT OUTER JOIN  dashboard AS folder ON dashboard.org_id = folder.org_id AND dashboard.folder_id = folder.id  WHERE ((dashboard.uid IN (SELECT substr(scope, 16) FROM permission WHERE scope LIKE 'dashboards:uid:%' AND role_id IN(SELECT id FROM role INNER JOIN (
                        SELECT ur.role_id
                        FROM user_role AS ur
                        WHERE ur.user_id = 1
                        AND (ur.org_id = 1 OR ur.org_id = 0)
                UNION
                        SELECT tr.role_id FROM team_role as tr
                        WHERE tr.team_id IN(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50)
                        AND tr.org_id = 1
                UNION
                        SELECT br.role_id FROM builtin_role AS br
                        WHERE br.role IN ("")
                        AND (br.org_id = 1 OR br.org_id = 0)
                ) as all_role ON role.id = all_role.role_id)  AND action = "dashboards:read") AND NOT dashboard.is_folder) OR (folder.uid IN (SELECT substr(scope, 13) FROM permission WHERE scope LIKE 'folders:uid:%'  AND role_id IN(SELECT id FROM role INNER JOIN (
                        SELECT ur.role_id
                        FROM user_role AS ur
                        WHERE ur.user_id = 1
                        AND (ur.org_id = 1 OR ur.org_id = 0)
                UNION
                        SELECT tr.role_id FROM team_role as tr
                        WHERE tr.team_id IN(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50)
                        AND tr.org_id = 1
                UNION
                        SELECT br.role_id FROM builtin_role AS br
                        WHERE br.role IN ("")
                        AND (br.org_id = 1 OR br.org_id = 0)
                ) as all_role ON role.id = all_role.role_id)  AND action = "dashboards:read") AND NOT dashboard.is_folder)) AND dashboard.org_id=1 AND dashboard.title LIKE "%dashboard_0_0%" AND dashboard.is_folder = 0 ORDER BY dashboard.title ASC LIMIT 1000 OFFSET 0) AS ids
                INNER JOIN dashboard ON ids.id = dashboard.id
LEFT OUTER JOIN dashboard AS folder ON folder.id = dashboard.folder_id
                LEFT OUTER JOIN dashboard_tag ON dashboard.id = dashboard_tag.dashboard_id
 ORDER BY dashboard.title ASC
```

</p>
</details> 

The diff output of the two looks like:
```
diff sql/old.sql sql/new.sql 
11c11
<                         folder.title AS folder_title  FROM ( SELECT dashboard.id FROM dashboard WHERE ((dashboard.uid IN (SELECT substr(scope, 16) FROM permission WHERE scope LIKE 'dashboards:uid:%' AND role_id IN(SELECT id FROM role INNER JOIN (
---
>                         folder.title AS folder_title  FROM ( SELECT dashboard.id FROM dashboard LEFT OUTER JOIN  dashboard AS folder ON dashboard.org_id = folder.org_id AND dashboard.folder_id = folder.id  WHERE ((dashboard.uid IN (SELECT substr(scope, 16) FROM permission WHERE scope LIKE 'dashboards:uid:%' AND role_id IN(SELECT id FROM role INNER JOIN (
24c24
<                 ) as all_role ON role.id = all_role.role_id)  AND action = "dashboards:read") AND NOT dashboard.is_folder) OR (dashboard.folder_id IN (SELECT d.id FROM dashboard as d WHERE d.uid IN (SELECT substr(scope, 13) FROM permission WHERE scope LIKE 'folders:uid:%'  AND role_id IN(SELECT id FROM role INNER JOIN (
---
>                 ) as all_role ON role.id = all_role.role_id)  AND action = "dashboards:read") AND NOT dashboard.is_folder) OR (folder.uid IN (SELECT substr(scope, 13) FROM permission WHERE scope LIKE 'folders:uid:%'  AND role_id IN(SELECT id FROM role INNER JOIN (
37c37
<                 ) as all_role ON role.id = all_role.role_id)  AND action = "dashboards:read")) AND NOT dashboard.is_folder)) AND dashboard.org_id=1 AND dashboard.title LIKE "%dashboard_0_0%" AND dashboard.is_folder = 0 ORDER BY dashboard.title ASC LIMIT 1000 OFFSET 0) AS ids
---
>                 ) as all_role ON role.id = all_role.role_id)  AND action = "dashboards:read") AND NOT dashboard.is_folder)) AND dashboard.org_id=1 AND dashboard.title LIKE "%dashboard_0_0%" AND dashboard.is_folder = 0 ORDER BY dashboard.title ASC LIMIT 1000 OFFSET 0) AS ids
```

**Why do we need this feature?**

[Add a description of the problem the feature is trying to solve.]

**Who is this feature for?**

[Add information on what kind of user the feature is for.]

**Which issue(s) does this PR fix?**:

<!--

- Automatically closes linked issue when the Pull Request is merged.

Usage: "Fixes #<issue number>", or "Fixes (paste link of issue)"

-->

Fixes #

**Special notes for your reviewer:**
I would suggest for reviewing to compare the `buildClauses()` implementation in the [existing](https://github.com/grafana/grafana/blob/f9d55c74ab1504d352d56549b74563e14eb0037d/pkg/services/sqlstore/permissions/dashboard.go#L113) and the [new](https://github.com/grafana/grafana/blob/f9d55c74ab1504d352d56549b74563e14eb0037d/pkg/services/sqlstore/permissions/dashboard_filter_no_subquery.go#L27) filter

I have run the benchmarks [before](https://drone.grafana.net/grafana/grafana/128770) and [after](https://drone.grafana.net/grafana/grafana/128789) [this](https://github.com/grafana/grafana/pull/72388/commits/f9d55c74ab1504d352d56549b74563e14eb0037d) commit that enables the newly introduced permission filter with the present changes - note that the MySQL 5.7 benchmarks timeout before completing without the present changes while with the changes are completed in less than 10 minutes - and the attach the `benchstat` results:

<details><summary>MySQL 5.7</summary>
<p>

```
% benchstat logs_grafana_grafana_128770_1_14.log logs_grafana_grafana_128789_1_14.log 
goos: linux
goarch: amd64
pkg: github.com/grafana/grafana/pkg/api
cpu: Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz
                                                                                         │ logs_grafana_grafana_128770_1_14.log │ logs_grafana_grafana_128789_1_14.log │
                                                                                         │                sec/op                │   sec/op     vs base                 │
FolderListAndSearch/get_root_folders_with_nested_folders_feature_enabled-32                                      101.34m ± 4%     99.63m ± 3%   -1.69% (p=0.038 n=8)
FolderListAndSearch/get_subfolders_with_nested_folders_feature_enabled-32                                         10.08m ± 6%     10.04m ± 7%        ~ (p=0.442 n=8)
FolderListAndSearch/list_all_inherited_dashboards_with_nested_folders_feature_enabled-32                          95.280 ± 1%      4.990 ± 2%  -94.76% (p=0.000 n=8)
FolderListAndSearch/search_for_pattern_with_nested_folders_feature_enabled-32                                     93.515 ± 0%      1.785 ± 5%  -98.09% (p=0.000 n=8)
FolderListAndSearch/search_for_specific_dashboard_nested_folders_feature_enabled-32                               93.968 ± 1%      1.812 ± 3%  -98.07% (p=0.000 n=8)
FolderListAndSearch/get_root_folders_with_nested_folders_feature_disabled-32                                      8.211m ± 0%     8.511m ± 3%   +3.65% (p=0.000 n=8)
FolderListAndSearch/list_all_dashboards_with_nested_folders_feature_disabled-32                                   107.61 ± 0%      20.30 ± 2%  -81.14% (p=0.000 n=8)
FolderListAndSearch/search_specific_dashboard_with_nested_folders_feature_disabled-32                            108.147 ±  ∞ ¹    1.401 ± 2%  -98.70% (p=0.044 n=2+8)
geomean                                                                                                            4.112          500.0m       -87.84%
¹ need >= 6 samples for confidence interval at level 0.95

                                                                                         │ logs_grafana_grafana_128770_1_14.log │  logs_grafana_grafana_128789_1_14.log  │
                                                                                         │                 B/op                 │     B/op       vs base                 │
FolderListAndSearch/get_root_folders_with_nested_folders_feature_enabled-32                                     5.901Mi ± 10%     6.173Mi ± 12%        ~ (p=0.279 n=8)
FolderListAndSearch/get_subfolders_with_nested_folders_feature_enabled-32                                       560.4Ki ± 39%     560.9Ki ± 48%        ~ (p=0.524 n=8)
FolderListAndSearch/list_all_inherited_dashboards_with_nested_folders_feature_enabled-32                        24.85Mi ±  8%     26.66Mi ± 15%        ~ (p=0.878 n=8)
FolderListAndSearch/search_for_pattern_with_nested_folders_feature_enabled-32                                   806.0Ki ±  7%     581.4Ki ± 10%  -27.88% (p=0.000 n=8)
FolderListAndSearch/search_for_specific_dashboard_nested_folders_feature_enabled-32                             553.7Ki ±  4%     293.3Ki ±  1%  -47.03% (p=0.000 n=8)
FolderListAndSearch/get_root_folders_with_nested_folders_feature_disabled-32                                    1.092Mi ±  0%     1.093Mi ±  0%        ~ (p=0.195 n=8)
FolderListAndSearch/list_all_dashboards_with_nested_folders_feature_disabled-32                                 26.70Mi ±  0%     26.33Mi ± 15%   -1.38% (p=0.000 n=8)
FolderListAndSearch/search_specific_dashboard_with_nested_folders_feature_disabled-32                          551.02Ki ±   ∞ ¹   69.93Ki ±  4%  -87.31% (p=0.044 n=2+8)
geomean                                                                                                         2.194Mi           1.522Mi        -30.61%
¹ need >= 6 samples for confidence interval at level 0.95

                                                                                         │ logs_grafana_grafana_128770_1_14.log │ logs_grafana_grafana_128789_1_14.log │
                                                                                         │              allocs/op               │  allocs/op   vs base                 │
FolderListAndSearch/get_root_folders_with_nested_folders_feature_enabled-32                                       152.4k ± 0%     152.4k ± 0%        ~ (p=0.741 n=8)
FolderListAndSearch/get_subfolders_with_nested_folders_feature_enabled-32                                         15.65k ± 0%     15.65k ± 0%        ~ (p=0.717 n=8)
FolderListAndSearch/list_all_inherited_dashboards_with_nested_folders_feature_enabled-32                          594.8k ± 0%     590.9k ± 0%   -0.66% (p=0.000 n=8)
FolderListAndSearch/search_for_pattern_with_nested_folders_feature_enabled-32                                    11.755k ± 3%     7.081k ± 0%  -39.76% (p=0.000 n=8)
FolderListAndSearch/search_for_specific_dashboard_nested_folders_feature_enabled-32                               5991.5 ± 6%      563.0 ± 3%  -90.60% (p=0.000 n=8)
FolderListAndSearch/get_root_folders_with_nested_folders_feature_disabled-32                                      24.58k ± 0%     24.58k ± 0%        ~ (p=1.000 n=8)
FolderListAndSearch/list_all_dashboards_with_nested_folders_feature_disabled-32                                   600.2k ± 0%     592.4k ± 0%   -1.31% (p=0.000 n=8)
FolderListAndSearch/search_specific_dashboard_with_nested_folders_feature_disabled-32                            10663.5 ±  ∞ ¹    466.0 ± 6%  -95.63% (p=0.044 n=2+8)
geomean                                                                                                           44.62k          21.02k       -52.89%
¹ need >= 6 samples for confidence interval at level 0.95

```

</p>
</details> 


<details><summary>MySQL 8.0</summary>
<p>

```
% benchstat logs_grafana_grafana_128770_1_15.log logs_grafana_grafana_128789_1_15.log 
goos: linux
goarch: amd64
pkg: github.com/grafana/grafana/pkg/api
cpu: Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz
                                                                                         │ logs_grafana_grafana_128770_1_15.log │ logs_grafana_grafana_128789_1_15.log │
                                                                                         │                sec/op                │    sec/op      vs base               │
FolderListAndSearch/get_root_folders_with_nested_folders_feature_enabled-32                                         104.4m ± 6%    103.7m ±  2%        ~ (p=0.279 n=8)
FolderListAndSearch/get_subfolders_with_nested_folders_feature_enabled-32                                           10.06m ± 8%    10.45m ±  6%        ~ (p=0.195 n=8)
FolderListAndSearch/list_all_inherited_dashboards_with_nested_folders_feature_enabled-32                            544.4m ± 8%   2355.1m ± 79%        ~ (p=0.083 n=8)
FolderListAndSearch/search_for_pattern_with_nested_folders_feature_enabled-32                                        8.078 ± 4%     4.671 ±  6%  -42.18% (p=0.000 n=8)
FolderListAndSearch/search_for_specific_dashboard_nested_folders_feature_enabled-32                                  7.621 ± 1%     4.133 ±  5%  -45.77% (p=0.000 n=8)
FolderListAndSearch/get_root_folders_with_nested_folders_feature_disabled-32                                        8.092m ± 5%    7.899m ±  2%   -2.39% (p=0.015 n=8)
FolderListAndSearch/list_all_dashboards_with_nested_folders_feature_disabled-32                                     510.7m ± 2%    345.2m ±  4%  -32.40% (p=0.000 n=8)
FolderListAndSearch/search_specific_dashboard_with_nested_folders_feature_disabled-32                              148.280 ± 0%     3.917 ±  4%  -97.36% (p=0.000 n=8)
geomean                                                                                                             619.1m         389.1m        -37.14%

                                                                                         │ logs_grafana_grafana_128770_1_15.log │ logs_grafana_grafana_128789_1_15.log │
…